Fixed issue where booleans were not being rendered (#3115)
