+++ app/src/componentsV2/CodeEditor/components/EditorV2/components/VariablePopOver/variable-popover.scss
